% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{canvasXpress.data}
\alias{canvasXpress.data}
\alias{canvasXpress.data-package}
\title{canvasXpress datasets}
\description{
This package contains data used and referenced in the canvasXpress R package
}
