% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get data for given filename.}
\usage{
get_data(file_name)
}
\arguments{
\item{file_name}{the file_name}
}
\value{
File content.
}
\description{
Get data for given filename.
}
