% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750}
\alias{GSE9750}
\title{Gene Expression Dataset - Cervical Cancer}
\format{A list of 3 dataframes containing prepared expression, gene, and sample data}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE9750}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GPL96}
}
\usage{
GSE9750
}
\description{
This example dataset is taken from the below referenced study.  That study was aimed 
at identification of gene expression profiles in cervical cancer and the role of specific genes
in cervical carcinogenesis.
}
\references{
Scotto L, Narayan G, Nandula SV, Arias-Pulido H et al. Identification of copy number 
gain and overexpressed genes on chromosome arm 20q by an integrative genomic approach in cervical 
cancer: potential role in progression. Genes Chromosomes Cancer 2008 Sep;47(9):755-65. PMID: 18506748

[HG-U133A] Affymetrix Human Genome U133A Array
}
\keyword{datasets}
