% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750_cleaned}
\alias{GSE9750_cleaned}
\title{Gene Expression Dataset - Cervical Cancer (cleaned)}
\format{A list of 3 dataframes containing prepared sample, expression, and gene data}
\usage{
GSE9750_cleaned
}
\description{
This example dataset is taken from the below referenced study.  That study was aimed 
at identification of gene expression profiles in cervical cancer and the role of specific genes
in cervical carcinogenesis.  The data was prepared for visualization by splitting the file into metadata (header)
information and gene expression values.  The data was cleaned - sample characteristics were divided to clean up the
metadata, missing values were encoded as NA, etc.  Gene information was brought in using the GEOquery package 
as a third dataframe to provide metadata on genes to allow data users to reference genes by Symbol, etc.
}
\references{
Scotto L, Narayan G, Nandula SV, Arias-Pulido H et al. Identification of copy number 
gain and overexpressed genes on chromosome arm 20q by an integrative genomic approach in cervical 
cancer: potential role in progression. Genes Chromosomes Cancer 2008 Sep; 47(9):755-65. PMID: 18506748

Davis, S. and Meltzer, P. S. GEOquery: a bridge between the Gene Expression Omnibus (GEO) and
BioConductor. Bioinformatics, 2007, 14, 1846-1847
}
\keyword{datasets}
