% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750_expression}
\alias{GSE9750_expression}
\title{Gene Expression Dataset - Cervical Cancer (expression values)}
\format{
A numeric matrix containing gene expression data
}
\usage{
GSE9750_expression
}
\description{
This example dataset is taken from the below referenced study.  That study was aimed 
at identification of gene expression profiles in cervical cancer and the role of specific genes
in cervical carcinogenesis.  

This dataset was prepared for visualization by retrieving the gene expression values that follow the source 
data header information and moving the ID column to the rownames.  This allows the storage and processing of the 
large numeric data as a matrix instead of a data frame for efficiency while retaining the ability to join it to
the sample annotations and gene information.
}
\references{
Scotto L, Narayan G, Nandula SV, Arias-Pulido H et al. Identification of copy number 
gain and overexpressed genes on chromosome arm 20q by an integrative genomic approach in cervical 
cancer: potential role in progression. Genes Chromosomes Cancer 2008 Sep; 47(9):755-65. PMID: 18506748
}
\keyword{datasets}
