% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750_gene_details}
\alias{GSE9750_gene_details}
\title{Gene Expression Dataset - Cervical Cancer (gene details)}
\format{A dataframe containing gene metadata for the GSE9750 study}
\usage{
GSE9750_gene_details
}
\description{
This example dataset is the gene information for genes present in the GSE9750 study referenced 
below.  The data was retrieved from the GEOquery package referenced below.

This dataset was prepared for visualization by renaming the columns and subsetting to only the information
necessary for visualization.  The Symbol and Description fields were also cleaned.
}
\references{
Davis, S. and Meltzer, P. S. GEOquery: a bridge between the Gene Expression Omnibus (GEO) and
BioConductor. Bioinformatics, 2007, 14, 1846-1847

Scotto L, Narayan G, Nandula SV, Arias-Pulido H et al. Identification of copy number 
gain and overexpressed genes on chromosome arm 20q by an integrative genomic approach in cervical 
cancer: potential role in progression. Genes Chromosomes Cancer 2008 Sep; 47(9):755-65. PMID: 18506748
}
\keyword{datasets}
