% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750_original}
\alias{GSE9750_original}
\title{Gene Expression Dataset - Cervical Cancer (original data)}
\format{}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE9750}
}
\usage{
GSE9750_original
}
\description{
This is the original, un-manipulated data from the below referenced study.  
That study was aimed  at identification of gene expression profiles in cervical cancer and the 
role of specific genes in cervical carcinogenesis.
}
\references{
Scotto L, Narayan G, Nandula SV, Arias-Pulido H et al. Identification of copy number 
gain and overexpressed genes on chromosome arm 20q by an integrative genomic approach in cervical 
cancer: potential role in progression. Genes Chromosomes Cancer 2008 Sep; 47(9):755-65. PMID: 18506748
}
\keyword{datasets}
