% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_data}
\alias{get_data}
\title{Get the file handle for given file in the package}
\usage{
get_data(file_name)
}
\arguments{
\item{file_name}{the file name to retrieve (from the extdata folder)}
}
\value{
File handle to the given file or "" if it does not exist
}
\description{
Get the file handle for given file in the package
}
\examples{

sannot_file <- get_data("network_compact.tsv")

# then read the file, for example:
# data <- read.delim(sannot_file)
# head(data)

}
