% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE9750}
\alias{GSE9750}
\title{Identification of gene expression profiles in cervical cancer}
\format{A list of 3 dataframes}
\usage{
GSE9750
}
\description{
Identification of gene expression profiles in cervical cancer
}
\keyword{datasets}
